Display both net score and raw components
In addition to displaying the net score, we want to display the number
of up/down votes so users can get a better idea of whether a link is
controversial or just boring.

Technical note: Since we don’t need to persist to the database and don’t
care about validity, we are using `Link.new` here instead of
`FactoryBot`.